finance-ai-pathway2/requirements.txt
# Core dependencies
pathway>=0.10.0
openai>=1.0.0
langchain>=0.1.0
llama-index>=0.9.0

# Web frameworks
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Financial data
yfinance>=0.2.0
feedparser>=6.0.0

# Environment and utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.0.0
flake8>=6.0.0

# Optional: For enhanced features
plotly>=5.17.0
streamlit-extras>=0.3.0